refactors companies_controller and improves responses

+++ psycopg/controllers/companies_controller.py
@@ -36,8 +36,8 @@ def company_create(request):
 
         return jsonify({"message": f"{company_name} has been added to the Companies table."}), 200
 
-    except:
-        cursor.rollback()
+    except Exception as e:
+        print(e)
         return jsonify({"message": "Company could not be added"}), 404
 
 
@@ -49,12 +49,23 @@ def companies_get():
             FROM Companies;
             """
         )
-        result = cursor.fetchall()
-        if result:
-            return jsonify(({"message": "companies found", "result": result})), 200
+        results = cursor.fetchall()
+        if results:
+            company_list = []
+
+            for company in results:
+                company_record = {
+                    'company_id': company[0],
+                    'company_name': company[1]
+                }
+
+                company_list.append(company_record)
+
+            return jsonify(({"message": "companies found", "result": company_list})), 200
         else:
             return jsonify(({"message": f"No companies found"})), 404
-    except:
+    except Exception as e:
+        print(e)
         return jsonify({"message": "Could not fetch companies data"}), 404
 
 
@@ -68,11 +79,22 @@ def company_get_by_id(company_id):
             """, [company_id]
         )
         result = cursor.fetchone()
+
         if result:
-            return jsonify(({"message": f"company with id {company_id} found", "result": result})), 200
+            company_list = []
+
+            company_record = {
+                'company_id': result[0],
+                'company_name': result[1]
+            }
+
+            company_list.append(company_record)
+
+            return jsonify(({"message": f"company with id {company_id} found", "result": company_list})), 200
         else:
             return jsonify(({"message": "No company found"})), 404
-    except:
+    except Exception as e:
+        print(e)
         return jsonify({"message": "Could not fetch company data"}), 404
 
 
@@ -91,27 +113,63 @@ def company_update_by_id(request, company_id):
                        [company_name, company_id])
         conn.commit()
 
-    except:
-        return jsonify({"message": "Product has been updated"}), 404
+        cursor.execute(
+            """
+            SELECT *
+            FROM Companies
+            WHERE company_id=%s;
+            """, [company_id]
+        )
+        result = cursor.fetchone()
 
-    return jsonify({"message": f"{company_name} has been updated."}), 200
+        if result:
+            company_list = []
+
+            company_record = {
+                'company_id': result[0],
+                'company_name': result[1]
+            }
+
+            company_list.append(company_record)
+
+            return jsonify(({"message": f"{company_name} has been updated", "result": company_list})), 200
+        else:
+            return jsonify(({"message": "No company found"})), 404
+
+    except Exception as e:
+        print(e)
+        return jsonify({"message": "Company could not be updated"}), 404
 
 
 def company_delete(company_id):
     company = {}
     company['company_id'] = int(company_id)
 
+    print('company: ', company)
     if not company['company_id']:
         return jsonify({"message": 'company_id is required'}), 400
 
     try:
-        cursor.execute("""
-            DELETE FROM Companies
-            WHERE company_id=%s
-            """, [company_id])
-        conn.commit()
+        cursor.execute(
+            """
+            SELECT *
+            FROM Companies
+            WHERE company_id=%s;
+            """, [company_id]
+        )
+        result = cursor.fetchone()
 
-    except:
+        if result:
+            cursor.execute("""
+                DELETE FROM Companies
+                WHERE company_id=%s
+                """, [company_id])
+            conn.commit()
+
+            return jsonify({"message": f"company with id {company_id} has been deleted."}), 200
+        else:
+            return jsonify({"message": f"no company exists with id {company_id}"}), 404
+
+    except Exception as e:
+        print(e)
         return jsonify({"message": "company could not be deleted"}), 404
-
-    return jsonify({"message": f"company with id {company_id} has been deleted."}), 200
